Replace arguments to `pasfmt_config!` with generic `:meta` arguments
This means that the user can freely specify whatever attributes they
want and they will be applied to the generated struct.

For example
```rust
pasfmt_config!(
  #[command(name = "foo")]
  #[command(version = "1.0")]
  #[command(author = "me")]
  Config
)
```

will set the name to "foo", the version to "1.0", and the author name to
"me", all without committing to any kind of interface in the macro.