Reuse allocations for input and output in `file_formatter`
In our benchmarks this amounts to about a 5% overall performance
improvement on my machine.

The only downside to this is that memory allocated for large files will
be held onto for longer than necessary. Given that memory usage is
pretty low overall (aligning pretty closely with the size of files
used), and speed is a higher priority for us right now than memory
usage, I think this is a worthwhile tradeoff.