Fix failing tests in CI for windows
Sometimes in CI the file timestamps don't update after a write to the
file. We don't know why exactly, but it probably has something to do
with their virtualisation or some deduping filesystem.

e.g. https://github.com/integrated-application-development/pasfmt/actions/runs/10590530822/job/29346465816#step:4:1159

Whatever the cause, it's clear we can't rely on timestamps to tell us
when files are touched; a more robust method is to lock the file and
assert that the operation fails as a result.