Restructure error handling in front-end and orchestrator crates
The inspiration for this refactoring was the fact that the `find_files`
function from `file_formatter.rs` was ignoring some errors and not
propagating the error to the caller. Specifically, it was ignoring IO
errors in enumerating directories, and glob errors did not result in a
non-zero exit code from `main`.

The new model for error handling all throughout the orchestrator and
front-end crates is to pass a callback from `main` to consume errors
from anywhere else. It's a lot like just using `log::error!` instead of
creating and returning an `Error` value, but it keeps all the power and
benefits of `Error` types (backtraces, context, universality in Rust).

This new model was chosen for a few reasons
1. It allows for errors to be logged (in a centralised way) as they
   happen, rather then after all the work is done.
2. It allows for errors to be (globally) non-terminating, which is
   usually what we want for the front-end of the formatter; each task
   should continue after any of the others have failed (unless they
   panicked).
3. It's not particularly useful at this stage to have any of these
   errors returned from the functions, because the calling code never
   wants to handle errors in any way other than to exit with a non-zero
   status later on. Also, returning the errors in the function
   signatures (e.g. via a `Vec`) would delay the logging, unless the
   logging is spread out to multiple other locations.
4. It prevents error messages from being logged multiple times. The
   previous approach would coalesce all the errors generated by the
   format into one 'summary' error, which would report how many errors
   were encountered and make sure the exit code is set properly. This
   wasn't very useful, and was redundant. It also created a stylistic
   clash with the error reporting, because the initial errors would be
   logged like 'ERROR ...' (with red colouring) whereas the final error
   would be returned from `main` and formatted like 'Error: ...'.