Switch from `figment` to `config` for configuration
They have similar feature sets, but figment has a few features we
don't need, avoiding which has lead to some awkward workarounds in our
code.

In particular, figment has a notion of provenance which enables better
error message in most cases by reporting the 'source' of the data. The
problem with this for our use-case, is when it comes to command-line
overrides. Figment would report that the source of these values was
`command_line.rs:<NUM>`, which is not very helpful for the user. We
worked around this with a wrapper type which overrode the source.

Figment also has a notion of 'profiles' which we also didn't care for.
It only served to make the error messages a bit more confusing for the
user by qualifying all the setting keys with a profile name like
'default'.

While we thought we had to work around the fact that the `Toml` provider
type would search for the file name in all parent directories, it was in
fact optional and we didn't realise it.

It was awkward, but possible, to programmatically provide overrides to
figment. The approach we used involved using the `Value::from_str`
method to parse the command-line value, which was necessary to allow
deserialising to any non-string type. This worked, but then made us opt
in to the escaping and trimming rules for this parser (e.g. '\n'), which
we don't really need.

The `config` crate on the other hand is much simpler for our use-case,
and has all the features we want, without many that we don't need. In
addition to this it's a more popular crate with associations with the
rustlang group, which makes it a more attractive option from a
maintenance point of view.

Finally, this switch cuts almost 400KiB off our final executable size,
which is nice.