Lazily initialise the windows default encoding
Placing the windows encoding behing a `LazyLock` ensures that the code
to call and process `GetACP` is called at most once.

After moving the call to `default_encoding` to inside the `serde`
default attribute this function should only be called at most once per
deserialised config object (i.e. effectively once per `pasfmt`
invocation). While this doesn't make too much of a difference at this
point, it could make a difference if multiple config objects were
created. Using `LazyLock` is so simple that it's hard to find a reason
not to do it.