Write formatted output to the console only as Unicode
With Rust on Windows, only UTF-8 data can be written to stdout when it's
connected to a Console. This is because internally Rust wants to use the
`WriteConsoleW` function from the winapi, which of course accepts UTF-16
data only. Rust checks that your data is valid UTF-8 and converts it to
UTF-16.

It doesn't make sense to use the original encoding when outputing the
formatted data to a console because
1. It's only for human consumption; the most important thing is that it
   is rendered properly, and Unicode is the best way to do that.
2. The data goes through winapi and the terminal emulator which will
   choose their own representation for the data; we can't know what
   encoding is used for data copied to the clipboard from the terminal.