Fix lexing of asm blocks with conditional `asm` keywords
A semi-common practice in some libraries with inline assembly is
wrapping the `asm` keyword in conditional directives. This usually means
that there are multiple of the `asm` keyword before any of the inline
assembly code actually starts.

```delphi
{$if A}
asm // maybe some other stuff here
{$else}
asm
{$endif}
end
```

The problem with our existing implementation is that it disables the
identification of keywords when 'inside' asm blocks, but the notion of
'inside' isn't quite correct because we lex before 'preprocessing'
(whereas the Delphi language is designed to be processed in a single
pass).

As an aside, the reason we try to not identify keywords inside asm
blocks is to prevent instructions with the same name as keywords (shl,
shr, xor) as being lexed as keywords. None of the other reserved words
can actually be used in inline assembly, so this seemed OK.