Tweak comment indentation
Before this change, the parser would default to setting the level of
comments to the level of the current context, making an exception for
comments inside a 'subroutine' or 'type block' context. This exception
allows some comments to inherit the level of the next line.

This change fixes an issue where only the last comment in a series
would have this exception applied, and also flips the default handling
so that this 'exception' is actually the norm.

Usually, comments are 'supposed' to relate to what they precede, so we
should try to set the level of the comment to the next 'thing'. However
in some cases this makes little sense, because the next 'thing' may be a
token that only exists to end a block context (e.g. 'end'). So the new
exception is for comments in 'statement list' contexts to have the level
of the context, but everything else has the level of the next line.