Add documentation for the configuration options
Initially I thought it would be enough to have the CLI spit out the
available options, but I realise now that it may be a key factor for a
user in deciding whether to download and run the tool at all.

It would be nice to have this generated to avoid the duplication, but
for now it's best to just quickly get some docs out.