Void totally ignored lines
A user might expect that they could use the `{pasfmt off}` comment to
get around a performance issue, or an error in one of the formatting
rules (e.g. 'no solution found'). However, since disabling formatting
works at the token level and not the line level, this isn't how it works
in practice.

This PR improves things by 'voiding' any line which consists of only
ignored tokens. Voiding the line means that none of the later formatting
rules will process it at all.