Deindent compiler directives in `package` files
Fixes #166.

Instead of *not* formatting the directives in dpk files, we instead
format them in a way consistent with what the IDE generates.

In the case of #202 we decided to format with a slightly different style
to the generated code for import clauses, because the difference was
only in a couple of rare cases. In this case, however, we have decided
to match the generated style because it affects 100% of dpk files, and
it's a pretty trivial difference.