Update keywords list in web demo
Fixes #208.

This change generates the list of keywords at build time from one of the
other scripts we have for keywords and codegen.

The differences between the old list and the new one are numerous, but
the short of it is that all Delphi keywords that the formatter itself
recognises are now recognised by the web demo.

I decided to include all the keywords, rather than just the reserved
words, because there are many important keywords that are not reserved
words.