Explore conditional directive paths more efficiently
This introduces a new approach for exploring the directive paths. This
new method involves repeatedly "passing" over the tree of conditional
directives, greedily taking the first non-exhausted branch at every
option (or the last if all are exhausted).

This approach is linear in the number of passes required, with respect
to the max conditional nesting level. The previous approach was
exponential.

This is a change in behaviour though; the resulting paths explored can
be different, which does result in a change in formatting for some code.
However, I've not yet seen a case where the formatting has changed from
something completely unbroken to something broken. Generally speaking,
for these cases that are right at the edge of what we can handle, the
formatting was and still is broken in some way.

Here's an example of something I've noticed in some libraries
```delphi
{$IFDEF CPU32}
procedure Foo;
{$ENDIF}
{$IFDEF CPU64}
procedure _Foo;
{$ENDIF}
begin
end;

procedure Bar;
begin
end;
```

This is hard to parse as a formatter because we would need to know that
the `CPU32` and `CPU64` branches are mutually exclusive. If we don't
know that, then there will be a pass of the file that looks like this
```delphi
procedure Foo;
  procedure _Foo;
  begin
  end;

  procedure Bar;
  begin
  end;
```

Where `_Foo` and `Bar` are subprocedures of `Foo`.

There are many other examples like that where the only way to correctly
parse the file is to understand the relationships between the
conditional directive branches.