Fix parsed level for comments outside statement lists
In cases of comments in place of statements (but not in directly in
statement *lists*), the parser would classify the comment as
"unfinished", which means that the level would be set to the level of
the next completed line.

This was mostly designed to detect comments that were at the top-level,
i.e. inside a type declaration, or before a routine declaration. It should
not be detecting comments that are merely in place of a statement.