Allow child lines of first token in line to be explored
In an earlier version of the patch for subroutines in anonymous
routines, I stumbled into an existing limitation of the optimising line
formatter. This limitation was that child lines of the very first token
in a logical line would never be explored, leading to zero indentation
for all such lines.

This oversight is a bit difficult to fix properly, because it is
reflected in some of the fundamental design of the algorithm. However,
it isn't too hard to hack it by treating the initial node as even *more*
of a special case than it already is.

As it stands, this change is effectively redundant, because the parser
still does not generate logical lines that relevant to this issue, but
it is entirely possible that it could generate such lines (either in a
future version, or in a fork / custom pasfmt version).