Convert anonymous routine section headers to child lines
Previously these headers would be in the parent line, and the lines
within the section would be children of the header (e.g.
var/const/type). With this change, every line in between the end of the
anonymous routine parameter list and the `begin` for the routine are
parsed as child lines of the initial `procedure` or `keyword` token.