Fix `Debug` implementation for `RawDebugLine`
PR #263 changed the way child lines map to their parent token. In the
case of an `if ... then {} ...;` construction, the mapped parent of the
statement went from `then` to `{}`. This meant that the child lines
would not be found or logged. This PR removes the special handling for
skipping comments to find child lines. This both simplifies the code
and fixes the bug.