Add version selector to web demo
This involves switching to dynamic loading of the WASM module, and also
involves changing how the GitHub pages is built and deployed.

The idea is that the latest version of the web demo site should be
backwards compatible with all the previous versions of the WASM module,
and the `gh-pages` branch will simply keep all the old WASM module
builds, while new ones are added and the web demo site is updated.

The default module version loaded is whatever is first in the
versions.json file that exists at the root of the `gh-pages` branch,
which the workflow makes sure to sort according to `sort -Vr` (version
sort in descending order). This means that the latest tagged version
should be first, and all the branches will be after all the tags.
- https://www.gnu.org/software/coreutils/manual/html_node/Version-sort-overview.html

In order to make switching between and comparing versions more
streamlined, I also implemented a function to migrate the settings as
you change the selected version. This comments and uncomments lines that
correspond to settings that were removed or added when moving between
versions.

The selected version is also a query parameter now, so old examples
shared via URL should remain stable as new versions are released.